flasktute/microservice/backend/requirements.txt
flask
pymongo
dnspython
python-dotenv
json



